backend/requirements.txt
numpy
Flask
pandas
matplotlib
scipy
redis
insightface
jupyter
pillow
opencv-python
scikit-learn